requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
PySide6>=6.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
